# internal/model/dto/paginatedPersonsDto.go
package dto

type PaginatedPersonsDto struct {
	Data       []PersonDto `json:"data"`
	Total      int         `json:"total"`
	Page       int         `json:"page"`
	PageSize   int         `json:"page_size"`
	TotalPages int         `json:"total_pages"`
}